Fix arguments provided to Godaddy lexicon provider
The arguments that the lexicon base provider accepts have changed for the update_record and delete_record methods. These now take an identifier as the first argument. The proposed changes here fix this issue and provide the required values as keyword arguments to prevent such argument order changes from breaking this plugin.